+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # edge-exchange-plugins
 
+## Unreleased
+
+- changed: Add `fantom` pluginId check to spookySwap and tombSwap
+
 ## 0.21.6 (2023-09-14)
 
 - changed: Move EVM data from spendTarget `otherParams` to `memo`